refactor(perf): split /creators + /magic into server shell + motion island
31st + 32nd in the pattern series.
